ai(danger): close round-two review — inheritance-resolve the cadence pairing, bound the evac pin from above, correct stale prose
INHERITANCE QUESTION — CLOSED, AND MORE STRONGLY THAN ADJACENCY SHOWED.
The two-sided cadence conclusion rested on "every weapon declaring ReloadDelay also
declares Magazine", checked by YAML adjacency. Resolved properly by walking each
weapon's Inherits chain across all 7 weapon files (161 entries: 128 concrete, 33
templates):

  - NO template declares Magazine or ReloadDelay at ALL. The only cadence fields any
    ^template carries are BurstWait/Burst (^TankRound, ^ArtilleryRound, ^TeslaWeapon,
    ^30mm.Tunguska). So inheritance CANNOT break the pairing — it never supplies
    either field.
  - resolved-only-through-inheritance = 0. Declared 14/14, resolved 14/14.
  - Violations (ReloadDelay > 0 with resolved Magazine <= 1) = 0.
  - Symmetric difference empty in both directions.
  - Magazine values: 4, 5x2, 10, 15, 20x4, 100x3, 300x2 — all > 1, range 4-300.

The pairing holds. The two-sided model stands. (Verification was a throwaway script,
not committed.)

FIX 1 — stale Stamp comments, and the correction the previous message owes.
The Stamp taper and accumulation comments still said "10^8" for a core intensity that
is now ~5.2e5, and Compute's said the rifleman stamped "~1,300" (that was the old
THROUGHPUT; the old intensity was ~1,626). Fixed, with real headroom stated: the taper
product is ~1.7e7 (~125x below int.MaxValue) and the accumulation would need ~4,000
co-located auras to saturate — the only term that scales with belief count rather than
the weapon table, which is why it stays saturating.

  CORRECTION TO d1591b35's COMMIT MESSAGE, since prose in a merged commit is immutable
  and this branch has already put one invented number into one: that message claimed
  "every comment that sold them as the fix now says otherwise". That was FALSE — the
  two Stamp comments still did. They do now.
  Also in that message: the doc worker's ~6,060 was said to "reconcile"; it does not
  quite — correct is 6,410. Direction and factor were right, the figure was not.

FIX 2 — the evac pin was one-sided.
It asserted only that a point-blank TANK (521,914) still fires against a threshold of
3,910, so EvacDangerUnits 50 -> 100, or even -> 500, left it green: it could catch a
threshold set too LOW (the shipped bug) but not one high enough to disable the evac.
Added the binding upper bound — a fully-believed RIFLEMAN on the truck's own cell
(7,820) must still fire — which narrows the passing band from ~[36, 6673] to ~[36, 99]
units. The test now also reads SupplyFollowerBotModuleInfo.EvacDangerUnits instead of
a literal 50, so retuning what ships moves the test. Verified RED at 2x AND at 0.5x.

FIX 3 — AdvanceDangerCeiling named in three [Desc] strings after the rename; these
ship into --docs as instructions naming a nonexistent field. Fixed. Its worked grid
was also stale ({20,35,50,65,80} -> {8,14,20,26,32} was computed against the old raw
20); against 6 danger units the same grid gives {0,0,6,12,18}, so the two cautious
points collapse and the dial loses low-end resolution. Documented rather than
re-tuned: OpportunisticAdvanceEnabled is default-off and unset in ai.yaml, and picking
a new slope wants the aggressiveness sweep it was originally measured with.

FIX 4 — ForwardStagingMath line refs, which shifted on THIS branch when the passable
parameter was added. :97-98/:116-117/:128-129 -> :112-113/:131-132/:150-151 in
PoiOffensiveBotModule's self-seed analysis and in DOCS/reference/influence-stack.md
(curated bank, fixed on sight per CLAUDE.md). Also noted at that call site that the new
passable filter would be a FOURTH way the walk returns its seed — the three-case
analysis stays exhaustive only because this caller passes no predicate.

NITs: ScoreDrop's two-channel sum widened to long (the header claimed the function
accumulates in long; it now does — not urgent, ~4,000 auras, but the claim should be
true); the saturating clamp in Compute is now exercised by a test asserting it stays
positive and monotone under absurd inputs, since an untested saturation path is
indistinguishable from a wrap until something reaches it; tab/space indentation fixed.

The DistributionLogEpisodes Desc no longer presents the 66,834 median as a calibration
— it was measured with the cadence bug live and describes a field that no longer
exists.

Build clean, 1275/1275. Unchanged from last round: the @stable benchmark baseline must
be re-taken, and the largest live risk is a real ruleset median lower than the fixture
estimate, which only a play log can answer — the instrumentation is there for it.

+++ engine/OpenRA.Mods.Common/Traits/BotModules/HelicopterSquadBotModule.cs
@@ -1813,11 +1813,17 @@ namespace OpenRA.Mods.Common.Traits
 			int enemyControlWeight, int dangerWeight, int reachWeight)
 		{
 			var enemyDepth = believedControl < 0 ? -believedControl : 0;
-			var danger = (groundDanger < 0 ? 0 : groundDanger) + (airDanger < 0 ? 0 : airDanger);
+
+			// Widened HERE, at the sum, not just at the multiply below: the two channels are independent
+			// readings and adding them in int would overflow before the weighting ever sees the value. It
+			// takes two ~1e9 auras to bite, which is far beyond the ~5e5 a single believed contact stamps —
+			// but the header below claims this whole function accumulates in long, and it should be true
+			// rather than nearly true.
+			var danger = (long)(groundDanger < 0 ? 0 : groundDanger) + (airDanger < 0 ? 0 : airDanger);
 			var reach = reachCells < 0 ? 0 : reachCells;
 
 			// ACCUMULATED IN long AND CLAMPED. `danger * dangerWeight` is a raw danger-field reading times a
-			// percentage — 10^7-10^8 times 100 — which overflows int and WRAPS NEGATIVE. Because every term
+			// percentage — 10^5-10^7 times 100 — which overflows int and WRAPS NEGATIVE. Because every term
 			// here is a penalty subtracted from 0, a wrapped danger term becomes a BONUS: the picker would
 			// actively prefer the most dangerous candidate it sampled, inverting the one lever this scorer
 			// exists to apply. Live under RiskWeightedDropSite.
